Move SelectViewModel out of Otlp namespace
