+++ src/Aspire.Dashboard/Components/Controls/SignalsActionsDisplay.razor.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Aspire.Dashboard.Model;
-using Aspire.Dashboard.Model.Otlp;
 using Aspire.Dashboard.Otlp.Storage;
 using Microsoft.AspNetCore.Components;
